refactor(HomeScreen): mover estilos para arquivo styles.js
