+++ mavis/src/HomeScreen/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import { StyleSheet } from 'react-native';
 import { Block, Text, Button, Icon } from 'galio-framework';
+import styles from './styles';
 
 export default function HomeScreen({ navigation }) {
   return (
@@ -105,66 +105,3 @@ export default function HomeScreen({ navigation }) {
     </Block>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-  },
-  content: {
-    flex: 1,
-    padding: 20,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  title: {
-    textAlign: 'center',
-    marginBottom: 20,
-    color: '#333',
-  },
-  description: {
-    textAlign: 'center',
-    marginBottom: 30,
-    color: '#555',
-    fontSize: 16,
-    lineHeight: 24,
-  },
-  menu: {
-    marginTop: 20,
-    width: '100%',
-    alignItems: 'center',
-  },
-  button: {
-    marginVertical: 10,
-    width: '100%',
-    backgroundColor: '#17a2b8',
-  },
-  emergencyButton: {
-    marginVertical: 20,
-    width: '100%',
-    backgroundColor: '#dc3545',
-  },
-  buttonContent: {
-    flexDirection: 'row',
-    alignItems: 'center',
-  },
-  buttonText: {
-    color: 'white',
-    marginLeft: 10,
-  },
-  bottomBar: {
-    height: 60,
-    flexDirection: 'row',
-    justifyContent: 'space-around',
-    alignItems: 'center',
-    backgroundColor: '#f8f9fa',
-    borderTopWidth: 1,
-    borderColor: '#ddd',
-  },
-  navButton: {
-    width: 50,
-    height: 50,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-});
